refactor: decompose SubscriptionServiceImpl into single-responsibility services
Split the 392-line god class into focused, testable services following
CQRS and existing gateway/resolver patterns:

- StripePaymentGateway: anti-corruption layer isolating all Stripe SDK calls
- SubscriptionQueryServiceImpl: read-only SubscriptionApi implementation
- SubscriptionCommandServiceImpl: user-facing commands (checkout, cancel, change plan)
- StripeWebhookHandlerImpl: webhook event processing
- SubscriptionLifecycleResolver: shared PlanType -> lifecycle template resolution
